REFACT: Further test division - split unit tests further; and - introduced a test utility class.
