+++ tests/utils.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+# coding=UTF-8
+#
+# E-ARK Validation
+# Copyright (C) 2019
+# All rights reserved.
+#
+# Licensed to the E-ARK project under one
+# or more contributor license agreements. See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership. The E-ARK project licenses
+# this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License. You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+"""Module that holds common utilities for unit testing."""
+from ip_validation.infopacks.rules import Severity
+
+def contains_rule_id(error_list, rule_id, severity=Severity.Error):
+    """Check that a particular error with specified severity is present in a list
+    of errors."""
+    for val_error in error_list:
+        if val_error.rule_id == rule_id:
+            if val_error.severity == severity:
+                return True
+    return False
